Add support for request ids
Request ids can be helpful in tracing and debugging requests when both
client and server log these. Inspiration is from
https://github.com/interagent/http-api-design#trace-requests-with-request-ids
and https://devcenter.heroku.com/articles/http-request-id.

The id of the request may come from an optional X-Request-ID header. The
value must be between 20 and 200 characters, and consist of ASCII letters,
digits, or the characters +, /, =, and -. Invalid or missing ids will be
ignored and replaced with generated ones.